Wrap tests as dummy bs modules
The `respect` unit test tool updated to use this package, and that gave
a circular package dependency.

However there is not actual circular dependency, because the tests in
this library depends on the library itself, and the `respect` library.
The `respect` library depends on this library. But the `respect`
library does not depend on the tests in this library.